+++ mono_core/lib/src/src.dart
@@ -18,6 +18,7 @@ export 'ports/package_scanner.dart';
 export 'ports/plugin_resolver.dart';
 export 'ports/prompt.dart';
 export 'ports/system.dart';
+export 'ports/logger_extensions.dart';
 export 'ports/target_selector.dart';
 export 'ports/task_executor.dart';
 export 'ports/task_plugin.dart';